requirements.txt
bokeh==3.6.1
Django==4.2.16
django-allauth==65.2.0
django-ckeditor-5==0.2.15
django-cookie-consent==0.6.0
django-ddm==2.1b1
django-qr-code==4.1.0
mysqlclient==2.2.6
numpy==2.1.3
pandas~=2.2.3
python-dateutil~=2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
wagtail==6.3.1
